Split server-payments.test.ts (1482→873 lines)
Moves GET /payment/success (ticket) and payment success token
verification tests into server-payments-success.test.ts (628 lines)
to bring both files under the 1200-line biome limit.
